removed spurious warning

+++ progs/cicada_extract_scfg_impl.hpp
@@ -527,7 +527,7 @@ struct ExtractSCFG
       const int source_count = alignment_count_source[iter->source.second] - alignment_count_source[iter->source.first];
       //const int target_count = alignment_count_target[iter->target.second] - alignment_count_target[iter->target.first];
       
-      const bool sentential_mode = sentential && source_length == source_size && target_length == target_size;
+      const bool sentential_mode = sentential && source_length == static_cast<int>(source_size) && target_length == static_cast<int>(target_size);
       const bool rule_mode = (! sentential) && (max_span <= 0 || source_length <= max_span);
       
       if (max_length <= 0 || source_length <= max_length || target_length <= max_length)
